ARMTest/requirements.txt
certifi-2023.11.17
charset-normalizer-3.3.2
contourpy-1.2.0
cycler-0.12.1
fonttools-4.46.0
idna-3.6
kiwisolver-1.4.5
lxml-4.9.3
matplotlib-3.8.2
numpy-1.26.2
packaging-23.2
pandas-2.1.4
pillow-10.1.0
pyparsing-3.1.1
python-dateutil-2.8.2
pytrends-4.9.2
pytz-2023.3.post1
requests-2.31.0
six-1.16.0
tzdata-2023.3
urllib3-2.1.0
